Bachelor Thesis Project using Python
Thesis Title : Text Encoding with Hybrid Scheme Using the Enigma Algorithm and the ElGamal Algorithm
In digital developments in the exchange of information, cryptography is still the method used to keep information (message) secret. The Enigma algorithm is one of symmetric cryptography. Since Enigma keys are susceptible to be hacked, it is necessary to increase message’s secrecy using hybrid cryptography methods. Hybrid cryptography is a combination of symmetric cryptography and asymmetric cryptography. One example of asymmetric cryptography is the ElGamal algorithm. In this research, a program is constructed to encode messages using hybrid cryptography between the Enigma algorithm and the ElGamal algorithm, which the Enigma key will be re-encrypted using the ElGamal algorithm to avoid messages being hacked by other parties. The ElGamal algorithm was chosen because of the difficulty of solving the discrete logarithm problem due the random number k chosen during the encryption process to be kept secret. The results of this research are the formation of hybrid scheme using the Enigma algorithm and the ElGamal algorithm as well as a application program made in the Python programming language by utilizing the Graphical User Interface (GUI) to make it easier for user (user-friendly).